backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# ─── Configuration ──────────────────────────────────────
pydantic-settings>=2.2.0

# ─── Supabase ───────────────────────────────────────────
supabase>=2.3.0

# ─── Computer Vision ────────────────────────────────────
opencv-python-headless>=4.9.0
mediapipe>=0.10.9

# ─── Machine Learning (CPU-only) ───────────────────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0

# ─── Numerical / Scientific ────────────────────────────
numpy>=1.26.0
scipy>=1.12.0

# ─── DTW ────────────────────────────────────────────────
dtw-python>=1.3.0

# ─── Testing ────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0
